% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceSummary.R
\name{add_legend}
\alias{add_legend}
\title{Add legend to the figure returned by \code{\link[mopa]{varianceSummary}}.}
\usage{
add_legend(...)
}
\arguments{
\item{...}{Named arguments to be passed to function \code{\link[graphics]{legend}}.}
}
\description{
Add legend to the figure returned by \code{\link[mopa]{varianceSummary}}.
}
\author{
M. Iturbide
}
