% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allModelling.R
\name{allModeling}
\alias{allModeling}
\title{Easy species distribution modelling and cross validation
into backgrounds of different extent}
\usage{
allModeling(data, varstack, k = 10, algorithm = c("glm", "svm", "maxent",
  "mars", "randomForest", "cart.rpart", "cart.tree"), destdir = getwd(),
  projection = CRS("+proj=longlat +init=epsg:4326"))
}
\arguments{
\item{data}{Object returned by function bindPresAbs or list/s of data frames with coordinates in the first two columns and presence/absence
(1=presence, 0=absence) in the third column.}

\item{varstack}{RasterStack of variables for modelling}

\item{k}{Integer. Number of folds for cross validation. Default is 10}

\item{algorithm}{Any character of the following: "glm", "svm", "maxent", "mars", "randomForest", "cart.rpart"
or "cart.tree"}

\item{destdir}{Character of the output path}

\item{projection}{Object of class CRS with the coordinate reference system. Default is
CRS("+proj=longlat +init=epsg:4326")}
}
\value{
Named Rdata objects are stored in the
specified path. Each Object is given a name indicating the algorithm, background
extent, and species in this order (if a single species is provided no name is given
for the species). The object returned by the function is a list of characters with listed files,
used algorithm, species names and background extents.
Each Rdata consists of a list with six components:

 \item{allmod }{fitted model using all data for training}
 \item{auc }{AUC statistic in the cross validation}
 \item{kappa }{kappa statistic in the cross validation}
 \item{tss }{true skill statistic in the cross validation }
 \item{mod }{fitted model with partitioned data}
 \item{p }{cross model prediction}
}
\description{
Species distribution modelling and k-fold cross validation
for a set of presence/absence data per species corresponding to a different background
extent. Algorithms supported are "glm", "svm", "maxent", "mars", "randomForest", "cart.rpart"
and "cart.tree"
}
\details{
This function calculates the AUC with the function \code{\link[PresenceAbsence]{auc}} from package
\pkg{PresenceAbsence}. \strong{Note:} Package \pkg{SDMTools} must be detached.
}
\examples{
\dontrun{
data(presaus)
data(biostackENSEMBLES)
##modeling
modirs <-allModeling(data = presaus, varstack = biostackENSEMBLES$baseline, k = 10, "mars")
}
}
\author{
M. Iturbide
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutierrez, J.M., 2015.
A framework for species distribution modelling with improved pseudo-absence generation. Ecological
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}

