% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{background}
\alias{background}
\title{Extract spatial background from raster}
\usage{
background(raster, projection = CRS("+proj=longlat +init=epsg:4326"))
}
\arguments{
\item{raster}{Raster object (typically corresponding to a variable used for modelling species distributions).}

\item{projection}{Object of class CRS with the coordinate reference system. Default is
CRS("+proj=longlat +init=epsg:4326")}
}
\value{
SpatialPoints object.
}
\description{
Obtain a spatial points object of the background given by a raster object.
}
\examples{
\dontrun{
data(biostack)
grid <- background(biostack$baseline$bio2)
plot(grid)
}
}
\author{
M. Iturbide
}

