% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bacgroundRadios.R
\name{backgroundRadios}
\alias{backgroundRadios}
\title{Background extent restriction for a sequence of distances}
\usage{
backgroundRadios(xy, background, start = 0.166, by = 0.083,
  unit = c("decimal degrees", "utm"))
}
\arguments{
\item{xy}{Data frame or list of data frames with coordinates (each row is a point)}

\item{background}{Matrix or list of matrixes of background coordinates.
Object derived from function \code{\link[mopa]{OCSVMprofiling}} ($absence).
Alternatively, object derived from function \code{\link[mopa]{delimit}} ($bbs.grid)
if the environmental profiling step is going to be avoided in the pseudo-absence
generation proccess).}

\item{start}{Value for the minimum distance to consider for extent limitations.
Default is 0.166}

\item{by}{Value of the distance to consider from one extent to the following.
Default is 0.083}

\item{unit}{Character indicating the coordinate system of the objects.
Default is "Decimal degrees", alternatively "utm" can be used}
}
\value{
List/s of matrixes with xy coordinates, each matrix correspond to a different
background extent.
}
\description{
Creation of point-grid backgrounds through the establishment of extent
limitations for a sequence of distances, with the length of the half diagonal of the
bounding box around xy records as the maximum distance
}
\details{
Argument unit is only used to set extent distances in km. This function forms part of the
second step in a three-step proccess to generate pseudo-absences, and is aimed at
creating backgrounds of different extent for pseudo-absence samplig based on an initial
point grid (derived from function \code{\link[mopa]{OCSVMprofiling}} or function
\code{\link[mopa]{backgroundGrid}}).
}
\examples{
\dontrun{
data(Oak_phylo2)
data(biostack)
projection(biostack$baseline) <- CRS("+proj=longlat +init=epsg:4326")
r <- biostack$baseline[[1]]
# Background of the whole study area
bg <- backgroundGrid(r)
## environmental profiling
bg.profiled <- OCSVMprofiling(xy = Oak_phylo2, varstack = biostack$baseline,
background = bg$xy)
bg.extents <- backgroundRadios(xy = Oak_phylo2, background = bg$xy,
start = 0.166, by = 0.083*10, unit = "decimal degrees")
bg.extents2 <- backgroundRadios(xy = Oak_phylo2, background = bg.profiled$absence,
start = 0.166, by = 0.083*10, unit = "decimal degrees")

}
}
\author{
M. Iturbide
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutierrez, J.M., 2015.
A framework for species distribution modelling with improved pseudo-absence generation. Ecological
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}
\seealso{
\code{\link[mopa]{mopaFitting}}, \code{\link[mopa]{pseudoAbsences}}, \code{\link[mopa]{backgroundGrid}},
\code{\link[mopa]{OCSVMprofiling}}
}

