% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bgRadio.R
\name{bgRadio}
\alias{bgRadio}
\title{Background extent restriction for a sequence of distances}
\usage{
bgRadio(xy, background, start = 0.166, by = 0.083,
  unit = c("decimal degrees", "utm"))
}
\arguments{
\item{xy}{Data frame or list of data frames with coordinates (each row is a point)}

\item{background}{Matrix or list of matrixes of background coordinates.
Object derived from function \code{\link[mopa]{OCSVMprofiling}} ($absence).
Alternatively, object derived from function \code{\link[mopa]{delimit}} ($bbs.grid)
if the environmental profiling step is going to be avoided in the pseudo-absence
generation proccess).}

\item{start}{Value for the minimum distance to consider for extent limitations.
Default is 0.166}

\item{by}{Value of the distance to consider from one extent to the following.
Default is 0.083}

\item{unit}{Character indicating the coordinate system of the objects.
Default is "Decimal degrees", alternatively "utm" can be used}
}
\value{
List/s of matrixes with xy coordinates, each matrix correspond to a different
background extent.
}
\description{
Creation of point-grid backgrounds through the establishment of extent
limitations for a sequence of distances, with the length of the half diagonal of the
bounding box around xy records as the maximum distance
}
\details{
Argument unit is only used to set extent names. This function forms part of the
second step in a three-step proccess to generate pseudo-absences, and is aimed at
creating backgrounds of different extent for pseudo-absence samplig based on an initial
point grid (derived from function \code{\link[mopa]{OCSVMprofiling}} or function
\code{\link[mopa]{delimit}}).
}
\examples{
\dontrun{
data(Oak_phylo2)
data(biostackENSEMBLES)
presences <- Oak_phylo2

##creation of point grid from raster object
sp_grid <- background(biostackENSEMBLES$baseline$bio2)

##delimit study area to the whole study domain for both species
bc <- rep(list(boundingCoords(coordinates(sp_grid))), length(presences))
del <- delimit(bounding.coords = bc, grid = sp_grid, names = names(presences))
## environmental profiling
unsuitable.bg <- OCSVMprofiling(xy = presences, varstack = biostackENSEMBLES$baseline,
bbs.grid = del$bbs.grid)
## sequence of 10 km between distances, from 20 km to the length of the
##half diagonal of the bounding box.
ext <- bgRadio(xy = presences,
background = unsuitable.bg$absence, start = 0.166,
by = 0.083, unit = "decimal degrees")

ext <- bgRadio(xy = presences,
background = del$bbs.grid, start = 0.166,
by = 0.083, unit = "decimal degrees")

# Plot presences for group H11 and background extents of 20, 120 and 520 km
plot(ext$H11$km520, col = "green4", pch = "*", asp = 1)
points(ext$H11$km120, pch = "*")
points(ext$H11$km20, pch = "*", col = "blue")
points(Oak_phylo2$H11, col = "red", pch = ".", cex = 1.5)
}
}
\author{
M. Iturbide
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutierrez, J.M., 2015.
A framework for species distribution modelling with improved pseudo-absence generation. Ecological
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}

