% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoAbsences.R
\name{bindPresAbs}
\alias{bindPresAbs}
\title{Bind presences and absences}
\usage{
bindPresAbs(presences, absences)
}
\arguments{
\item{presences}{Data frame or list of data frames with coordinates for presence data 
(each row is a point)}

\item{absences}{Object returned by function \code{\link[mopa]{pseudoAbsences}}. 
List/s of data frames with coordinates for absence data (each row is a point)}
}
\value{
List/s of matrixes with xy coordinates for presence/pseudo-absence data.
Each matrix correspond to a different background extent
}
\description{
Binds presence and absence data for each background extension
}
\author{
M. Iturbide \email{maibide@gmail.com}
}
