% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdata.R
\docType{data}
\name{biostack}
\alias{biostack}
\title{Bioclim RasterStack}
\description{
A RasterStack dataset containing three bioclimatic variables (bio1, bio13, bio14).
}
\author{
M. Iturbide
}
\keyword{variables}

