% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backgroundGrid.R
\name{boundingCoords}
\alias{boundingCoords}
\title{Bounding box coordinates of xy records}
\usage{
boundingCoords(xy)
}
\arguments{
\item{xy}{Data frame or list of data frames with coordinates (each row is a point)}
}
\value{
A vector or a list of vectors with bounding coordinates 
in the following form: c(x1, x2, y1, y2).
}
\description{
Creates a vector or a list of vectors of bounding coordinates around point records (xy records).
This is an internal function used by function
\code{\link{backgroundGrid}}.
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutierrez, J.M., 2015. 
A framework for species distribution modelling with improved pseudo-absence generation. Ecological 
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}
\author{
M. Iturbide
}
\keyword{internal}
