% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mopaTrain.R
\name{cutTSS}
\alias{cutTSS}
\title{Cut value of the max TSS}
\usage{
cutTSS(Obs, Fit)
}
\arguments{
\item{Obs}{Observed values}

\item{Fit}{fitted values}
}
\value{
threshold that of the max TSS
}
\description{
Cut value of the max TSS
}

