% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFromModel.R
\name{extractFromModel0}
\alias{extractFromModel0}
\title{Internal function to extrac objects from lists returned by function \code{\link[mopa]{mopaTrain}}}
\usage{
extractFromModel0(models, value = c("model", "auc", "kappa", "tss",
  "fold.models", "ObsPred"))
}
\arguments{
\item{models}{Object returned by \code{\link[mopa]{mopaTrain}}.}

\item{value}{Any character of the following: "model", "auc", "kappa", "tss", "fold.models", "ObsPred"}
}
\value{
Depending on the specified value:
 \item{model }{fitted model using all data for training}
 \item{auc }{AUC statistic in the cross validation}
 \item{kappa }{kappa statistic in the cross validation}
 \item{tss }{true skill statistic in the cross validation }
 \item{fold.models }{fitted model with partitioned data}
 \item{ObsPred }{Observed and prediced (cross model prediction) values}
}
\description{
Internal function to extract values returned by function \code{\link[mopa]{mopaTrain}}
}
\details{
\code{ObsPred} allows to calculate further accuracy measures.
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutierrez, J.M., 2015. 
A framework for species distribution modelling with improved pseudo-absence generation. Ecological 
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}
\author{
M. Iturbide
}
