% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kappaRepet.R
\name{kappaRepet}
\alias{kappaRepet}
\title{Cut value of the max TSS}
\usage{
kappaRepet(Obs, Fit, TSS = TRUE)
}
\arguments{
\item{Obs}{Observed values}

\item{Fit}{fitted values}

\item{TSS}{Default is TRUE}
}
\value{
threshold that of the max TSS
}
\description{
Cut value of the max TSS
}

