% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaveOneOut.R
\name{leaveOneOut}
\alias{leaveOneOut}
\title{Leave out a different subset for test each fold}
\usage{
leaveOneOut(kf.data)
}
\arguments{
\item{kf.data}{Object returned by function kfold or list of data frames}
}
\value{
List of length k. Each slot contains another list of two components:
 \item{train }{data frame of training data}
 \item{test }{data frame for test}
}
\description{
Given a data frame splitted in k subsets, this function binds 
subsets for training data and leaves one subset out for test k times, leaving 
out a different subset each fold.
}
\details{
Details.
}
\author{
M. Iturbide
}
\keyword{internal}
