% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadTestValues}
\alias{loadTestValues}
\title{Load model validation statistics from Rdata generated with function allModelling}
\usage{
loadTestValues(data, test = c("auc", "kappa", "tss"), algorithm = c("glm",
  "svm", "maxent", "mars", "randomForest", "cart.rpart", "cart.tree"),
  sourcedir = getwd())
}
\arguments{
\item{data}{Object with the same structure as the object returned by function bindPresAbs.}

\item{test}{Any character of the following: "auc", "kappa", "tss".}

\item{algorithm}{Any character of the following: "glm", "svm", "maxent", "mars", "randomForest",
"cart.rpart" or "cart.tree"}

\item{sourcedir}{Character of the path where Rdata objects are}
}
\value{
Matrix with values for each species and background extents.
}
\description{
Load model validation statistics from Rdata generated with function allModelling
}
\details{
detail.
}
\examples{
data(presaus)
data(biostack)
##modeling
modirs <-allModeling(data = presaus, varstack = biostack, k = 10, "mars")
##loading
auc_mars <-loadTestValues(data = presaus, test = "auc", algorithm = "mars")

library(lattice)
levelplot(auc_mars ,aspect=5 ,at =seq(0.5,1,0.01), col.regions=bpy.colors,
xlab="Haplogroups" ,ylab="Background extent (km)", main = "AUC")
}
\author{
M. Iturbide \email{maibide@gmail.com}
}

