% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_function.R
\name{mel_function}
\alias{mel_function}
\title{A Mel Function}
\usage{
mel_function(love = TRUE)
}
\arguments{
\item{love}{Do you love Mel?}
}
\value{
the answer/output of function
}
\description{
This function allows you to express your love of Mel
}
\examples{
mel_function()
}
\keyword{Mel}

