% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelo.R, R/mopaTrain.R
\name{modelo}
\alias{modelo}
\title{Species distribution modelling and cross validation}
\usage{
modelo(kdata, data, algorithm = c("glm", "svm", "maxent", "mars",
  "randomForest", "cart.rpart", "cart.tree"), weighting = FALSE,
  threshold = 0.5)

modelo(kdata, data, algorithm = c("glm", "svm", "maxent", "mars",
  "randomForest", "cart.rpart", "cart.tree"), weighting = FALSE,
  threshold = 0.5)
}
\arguments{
\item{kdata}{Object returned by function leaveOneOut}

\item{data}{Object returned by function biomat. 2D matrix with the dependent variable
(presence/absence) in the first column and the independent variables in the rest
(extracted from varstack)}

\item{algorithm}{Any character of the following: \code{"glm"}, "svm", "maxent", "mars", "randomForest", "cart.rpart"
or "cart.tree"}

\item{weighting}{Logical for model fitting with weighted presence/absences-s. Applicable for algorithms "glm", "mars",
"randomForest" and "cart.rpart". Default is FALSE.}

\item{threshold}{Cut value between 0 and 1 to calculate the confusion matrix. Default is 0.5.}

\item{kdata}{Object returned by function leaveOneOut}

\item{data}{Object returned by function biomat. 2D matrix with the dependent variable
(presence/absence) in the first column and the independent variables in the rest
(extracted from varstack)}

\item{algorithm}{Any character of the following: \code{"glm"}, "svm", "maxent", "mars", "randomForest", "cart.rpart"
or "cart.tree"}

\item{weighting}{Logical for model fitting with weighted presence/absences-s. Applicable for algorithms "glm", "mars",
"randomForest" and "cart.rpart". Default is FALSE.}

\item{threshold}{Cut value between 0 and 1 to calculate the confusion matrix. Default is 0.5.}
}
\value{
A list with six components:
 \item{allmod }{fitted model using all data for training}
 \item{auc }{AUC statistic in the cross validation}
 \item{kappa }{kappa statistic in the cross validation}
 \item{tss }{true skill statistic in the cross validation }
 \item{mod }{fitted model with partitioned data}
 \item{p }{cross model prediction}

A list with six components:
 \item{model }{fitted model using all data for training}
 \item{auc }{AUC statistic in the cross validation}
 \item{kappa }{kappa statistic in the cross validation}
 \item{tss }{true skill statistic in the cross validation }
 \item{fold.models }{fitted model with partitioned data}
 \item{ObsPred }{cross model prediction}
}
\description{
Species distribution modelling with k-fold cross validation.
Algorithms supported are "glm", "svm", "maxent", "mars", "randomForest", "cart.rpart"
and "cart.tree"

Species distribution modelling with k-fold cross validation.
Algorithms supported are "glm", "svm", "maxent", "mars", "randomForest", "cart.rpart"
and "cart.tree"
}
\details{
This function calculates the AUC with the function "auc" from package
"PresenceAbsence". Package SDMTools must be detached.

This function calculates the AUC with the function "auc" from package
"PresenceAbsence". Package SDMTools must be detached.
}
\author{
M. Iturbide

M. Iturbide
}
\keyword{internal}

