% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nicheOver.R
\name{nicheOver}
\alias{nicheOver}
\title{Niche overlap}
\usage{
nicheOver(stack, metric = c("D", "O"))
}
\arguments{
\item{stack}{RasterStack}

\item{metric}{Metric for niche overlap. Options are "D" and "O" (see details).}
}
\value{
Matrix of overlap values for metric D or O.
}
\description{
Compute niche overlap among rasters in a RasterStack
}
\details{
Niche overlap measures the similarity of the environmental ranges occupied by each constructed model
via operating the difference between two vectors of probability distributions (p), where p x,i and p y,i are
the normalized suitability scores for biological entity or model X and Y in grid cell i.

 D: Schoener’s statistic for niche overlap (Warren et al., 2008).
 O: Pianka index (Pianka, 1973).
}
\author{
M. Iturbide
}
\references{
Warren DL, Glor RE, Turelli M, Funk D (2008) Environmental niche equivalency versus conservatism:
Quantitative approaches to niche evolution. Evolution, 62, 2868–2883. doi:10.1111/j.1558-5646.2008. 00482.x.

Pianka ER (1973) The Structure of Lizard Communities. Annual Review of Ecology and Systematics, 4,
53–74. doi:10.1146/annurev.es.04.110173.000413.
}

