% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata.R
\docType{data}
\name{Oak_phylo2}
\alias{Oak_phylo2}
\title{Oak distribution}
\source{
Oak_phylo2 is a modified subset of the \strong{Quercus sp Europe Petit 2002} 
database (Petit et al., 2002b), which is available in the \strong{Georeferenced Database 
of Genetic Diversity} or \strong{(GD)^2}.
}
\description{
A dataset consisting of a list with two
data frames with xy coordinates corresponding to the distribution of two 
phylogenetic groups of oaks (H11 and H1)
}
\references{
Petit, R. J. \emph{et al} 2002. Chloroplast DNA variation in european 
white oaks: Phylogeography and patterns of diversity based on data from over 2600 populations. 
Forest Ecology and Management 156 (1-3), 5-26.

Evolution of Trees and Forest Communities \cr 
Ten years of the EVOLTREE network \cr
Evoltree E-Lab - An information system for forest genetics \cr
ISBN: 978-2-9519296-3-9
}
\keyword{records}
\keyword{xy}
