% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predictAllmod.R
\name{predictAllmod}
\alias{predictAllmod}
\title{Model prediction}
\usage{
predictAllmod(models, varstack)
}
\arguments{
\item{models}{model class object or list of model class objects, e.g. as returned by function `loadDefinitiveModel`.}

\item{varstack}{RasterStack of variables for projecting}
}
\value{
RasterStack of the projected probabilities
}
\description{
Model projection into a RasterStack
}
\examples{
\dontrun{
data(presaus)
data(biostackENSEMBLES)
##modeling
modirs <-allModeling(data = presaus, varstack = biostackENSEMBLES$baseline, k = 10, "mars")
##loading
auc_mars <-loadTestValues(models = modirs, test = "auc")
ind <- indextent(testmat = auc_mars, diagrams = F)
def <- loadDefinitiveModel(defExtents = ind, slot = "allmod")
##predicting
p <- predictAllmod(def, biostackENSEMBLES$future_71_100_SMHI_ECHAM5)
spplot(p$H11)
}
}
\author{
M. Iturbide
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutierrez, J.M., 2015.
A framework for species distribution modelling with improved pseudo-absence generation. Ecological
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}

