% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{presaus}
\alias{presaus}
\title{Presence/Pseudo-absence data for different background extents}
\description{
A dataset with listed data frames for two different Oak groups (H11 and H5).
Each data frame is the distribution of presences and pseudo-absences inside
a certain background extent
}
\author{
M. Iturbide \email{maibide@gmail.com}
}
\keyword{background}

