% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PseudoAbsences}
\alias{PseudoAbsences}
\title{Pseudo-absences}
\usage{
PseudoAbsences(xy, bg.grids, exclusion.buffer = 0.0166, prevalence = 0.5,
  kmeans = FALSE, projection = CRS("+proj=longlat +init=epsg:4326"))
}
\arguments{
\item{xy}{Data frame or list of data frames with coordinates (each row is a point)}

\item{bg.grids}{Object derived from function OCSVMprofiling or bgRadio. (List/s of matrixes
with background xy coordinates)}

\item{exclusion.buffer}{value of the minimum distance to be kept between presence data and
pseudo-absence data. Default is 0.0166}

\item{prevalence}{Proportion of presences against absences. Default is 0.5 (equal number of
presences and absences)}

\item{kmeans}{Logical. If FALSE (default) pseudo-absences are generated at random. If TRUE
k-means clustering of the background is done and centroids are extracted as pseudo-absences.}

\item{projection}{Object of class CRS with the coordinate reference system. Default is
CRS("+proj=longlat +init=epsg:4326")}
}
\value{
List/s of data frames
}
\description{
Pseudo-absence data generation at random or by k-means clustering inside backgrounds of
different extents
}
\details{
Details.The application of this function is the third step in a three-step proccess
to generate pseudo-absences.
}
\examples{
\donttest{
##delimit study area
data(Oak_phylo2)
data(sp_grid)
oak.extension<-boundingCoords(Oak_phylo2)
box.grid<-delimit(oak.extension, sp_grid, names(Oak_phylo2))
## environmental profiling
data(biostack)
unsuitable.bg <-OCSVMprofiling(xy = Oak_phylo2, varstack = biostack,
bbs.grid = box.grid$bbs.grid)
## sequence of 100 km between distances, from 20 km to the length of the
##half diagonal of the bounding box.
ext <-bgRadio(xy = Oak_phylo2, bounding.coords = oak.extension,
bg.absence = unsuitable.bg$absence, start = 0.166, by = 0.83, unit = "decimal degrees")
## pseudo-absence generation at random
pa_random <-PseudoAbsences(xy = Oak_phylo2, bg.grids = ext,
exclusion.buffer = 0.0083, prevalence = 0.5, kmeans = FALSE)
##plot
plot(ext$H11[[5]], pch="*", col= "grey", cex=.5)
points(pa_random$H11[[5]], col="red", pch=".", cex=4)
points(Oak_phylo2$H11, col="blue", pch=".", cex=3)
}
}
\author{
M. Iturbide \email{maibide@gmail.com}
}

