% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata.R
\docType{data}
\name{Q_pubescens}
\alias{Q_pubescens}
\title{Quercus pubsencens distribution}
\source{
Q_pubescens is a modified subset of occurrences obtained from \strong{GBIF.org}
}
\description{
A data frame with xy coordinates of Quercus pubesnces distribution
}
\references{
GBIF.org (14th March 2017) GBIF Occurrence Download http://doi.org/10.15468/dl.4ss6vr
}
\keyword{records}
\keyword{xy}
