% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_grid.R
\docType{data}
\name{sp_grid}
\alias{sp_grid}
\title{World grid}
\description{
A dataset of class SpatialPoints covering the world at 10km resolution
}
\author{
M. Iturbide \email{maibide@gmail.com}
}
\keyword{background}

