% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdata.R
\docType{data}
\name{sp_grid_europe}
\alias{sp_grid_europe}
\title{Europe grid}
\description{
A dataset of class SpatialPoints covering Europe at 30km resolution
}
\author{
M. Iturbide
}
\keyword{background}

