% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdata.R
\docType{data}
\name{sp_grid_world}
\alias{sp_grid_world}
\title{World grid}
\description{
A dataset of class SpatialPoints covering the world at 10km resolution
}
\author{
M. Iturbide
}
\keyword{background}

