% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mopaTrain.R
\name{TSS.Stat}
\alias{TSS.Stat}
\title{Internarl function for cutTSS}
\usage{
TSS.Stat(data)
}
\arguments{
\item{data}{data generated in function cutTSS}
}
\description{
Internarl function for cutTSS
}
\keyword{internal}
