% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/varianceAnalysis.R
\name{varianceAnalysis}
\alias{varianceAnalysis}
\title{Variance analysis of RasterStack objects}
\usage{
varianceAnalysis(predictions, component1, component2)
}
\arguments{
\item{predictions}{list of raster objects as returned by \code{\link[mopa]{mopaPredict}}}

\item{component1}{Character of the names in \code{predictions} that correspond to the first component in
the variance analysis.}

\item{component2}{Character of the names in \code{predictions} that correspond to the second component in
the variance analysis.}
}
\value{
A list of two RasterStack objects, the first containing the global mean and standard deviation and the
second containing the percentage of variance correponding to each component in the analysis
(component1, component2 and components 1 & 2).
}
\description{
Extract componets of list of objects
(as returned by function \code{\link[mopa]{mopaPredict}})
and perform variance analysis to obtain raster objects of the
contribution of each component to the observed varaibility.
}
\details{
Rasters are extracted from \code{predictions} using function \code{\link[base]{grep}}, by matching
names in \code{predictions} and characters in \code{componen1} and \code{componen2}.
The contribution of componen1 in front component2 to the spread (uncertainty) of the projected probabilities
in \code{predictions} is here assessed using a simple analysis of variance approach, where the total
variance (V) can be decomposed as the summation of the variance explained by component1 (Vcomp1),
component2 (Vcomp2) and the combination of the previous two (Vcomp12):

\eqn{V = Vcomp1 + Vcomp2 + Vcomp2}.
}
\examples{
\dontrun{
data(Oak_phylo2)
data(biostack)
projection(biostack$baseline) <- CRS("+proj=longlat +init=epsg:4326")
r <- biostack$baseline[[1]]

## Background of the whole study area
bg <- backgroundGrid(r)


## considering an unique background extent
RS_random <-pseudoAbsences(xy = Oak_phylo2, background = bg$xy,
realizations = 10,
exclusion.buffer = 0.083*5, prevalence = -0.5, kmeans = FALSE)

fittingRS <- mopaTrain(y = RS_random, x = biostack$baseline, k = 10,
algorithm = "glm", weighting = TRUE)

modsRS <- extractFromModel(models = fittingRS, value = "model")

#MODEL PREDICTION AND ANALYSIS OF THE VARIABILITY IN PROJECTIONS
prdRS.fut <- mopaPredict(models = modsRS, varstack = biostack$future)
component2 <- names(prdRS.fut$realization01)
component1 <- names(prdRS.fut)
result <- varianceAnalaysis(prdRS.fut, component1, component2)
spplot(result$variance, col.regions = rev(get_col_regions()))
}
}
\author{
M. Iturbide
}
\references{
San-Martín, D., Manzanas, R., Brands, S., Herrera, S. & Gutiérrez, J.M. (2016) Reassessing
Model Uncertainty for Regional Projections of Precipitation with an Ensemble of Statistical Downscaling Methods.
Journal of Climate 30, 203–223.
}

