% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bgRadio.R
\name{bgRadio}
\alias{bgRadio}
\title{Background extent restriction for a sequence of distances}
\usage{
bgRadio(xy, bounding.coords, bg.absence, start = 0.166, by = 0.083,
  unit = "decimal degrees")
}
\arguments{
\item{xy}{Data frame or list of data frames with coordinates (each row is a point)}

\item{bounding.coords}{Object returned by function \code{\link[mopa]{boundingCoords}}.
Matrix or list of matrixes of bounding coordinates indicating the maximum and minimum
values in columns and xy coordinates in rows}

\item{bg.absence}{Object derived from function \code{\link[mopa]{OCSVMprofiling}} ($absence).
Alternatively, object derived from function \code{\link[mopa]{delimit}} ($bbs.grid)
if the environmental profiling step is going to be avoided in the pseudo-absence
generation proccess).}

\item{start}{Value for the minimum distance to consider for extent limitations.
Default is 0.166}

\item{by}{Value of the distance to consider from one extent to the following.
Default is 0.083}

\item{unit}{Character indicating the coordinate system of the objects.
Default is "Decimal degrees", alternatively "utm" can be used}
}
\value{
List/s of matrixes with xy coordinates, each matrix correspond to a different
background extent.
}
\description{
Creation of point-grid backgrounds through the establishment of extent
limitations for a sequence of distances, with the length of the half diagonal of the
bounding box around xy records as the maximum distance
}
\details{
Argument unit is only used to set extent names. This function forms part of the
second step in a three-step proccess to generate pseudo-absences, and is aimed at
creating backgrounds of different extent for pseudo-absence samplig based on an initial
point grid (derived from function \code{\link[mopa]{OCSVMprofiling}} or function
\code{\link[mopa]{delimit}}).
}
\examples{
\dontrun{
##delimit study area
data(Oak_phylo2)
data(sp_grid)
oak.extension<-boundingCoords(Oak_phylo2)
box.grid<-delimit(oak.extension, sp_grid, names(Oak_phylo2))
## environmental profiling
data(biostack)
unsuitable.bg <-OCSVMprofiling(xy = Oak_phylo2, varstack = biostack,
bbs.grid = box.grid$bbs.grid)
## sequence of 10 km between distances, from 20 km to the length of the
##half diagonal of the bounding box.
ext <- bgRadio(xy = Oak_phylo2, bounding.coords = oak.extension,
bg.absence = unsuitable.bg$absence, start = 0.166,
by = 0.083, unit = "decimal degrees")

# Plot presences for group H11 and background extents of 20, 120 and 520 km
plot(ext$H11$km520, col = "green4", pch = "*", asp = 1)
points(ext$H11$km120, pch = "*")
points(ext$H11$km20, pch = "*", col = "blue")
points(Oak_phylo2$H11, col = "red", pch = ".", cex = 1.5)
}
}
\author{
M. Iturbide \email{maibide@gmail.com}
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutierrez, J.M., 2015.
A framework for species distribution modelling with improved pseudo-absence generation. Ecological
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}

