% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bindPresAbs.R
\name{bindPresAbs}
\alias{bindPresAbs}
\title{Bind presences and absences}
\usage{
bindPresAbs(presences, absences)
}
\arguments{
\item{presences}{Data frame or list of data frames with coordinates for presence data
(each row is a point)}

\item{absences}{Object returned by function \code{\link[mopa]{pseudoAbsences}}.
List/s of data frames with coordinates for absence data (each row is a point)}
}
\value{
List/s of matrixes with xy coordinates for presence/pseudo-absence data.
Each matrix correspond to a different background extent
}
\description{
Binds presence and absence data for each background extension
}
\examples{
\dontrun{
##delimit study area
data(Oak_phylo2)
data(sp_grid)
oak.extension<-boundingCoords(Oak_phylo2)
box.grid<-delimit(oak.extension, sp_grid, names(Oak_phylo2))
## environmental profiling
data(biostack)
unsuitable.bg <-OCSVMprofiling(xy = Oak_phylo2, varstack = biostack,
bbs.grid = box.grid$bbs.grid)
## sequence of 100 km between distances, from 20 km to the length of the
##half diagonal of the bounding box.
ext <-bgRadio(xy = Oak_phylo2, bounding.coords = oak.extension,
bg.absence = unsuitable.bg$absence, start = 0.166, by = 0.083, unit = "decimal degrees")
## pseudo-absence generation at random
pa_random <-PseudoAbsences(xy = Oak_phylo2, bg.grids = ext,
exclusion.buffer = 0.083, prevalence = 0.5, kmeans = FALSE)
presaus <-bindPresAbs(presences = Oak_phylo2, absences = pa_random)
}
}
\author{
M. Iturbide \email{maibide@gmail.com}
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutierrez, J.M., 2015.
A framework for species distribution modelling with improved pseudo-absence generation. Ecological
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}

