% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biomat.R
\name{biomat}
\alias{biomat}
\title{Matrix with variables for modelling}
\usage{
biomat(data, varstack)
}
\arguments{
\item{data}{Data frame with coordinates in the first two columns and presence/absence
(1=presence, 0=absence) in the third column.}

\item{varstack}{RasterStack of variables for modelling}

\item{projection}{Object of class CRS with the coordinate reference system. Default is
CRS("+proj=longlat +init=epsg:4326")}
}
\value{
2D matrix with the dependent variable (presence/absence) in the first column
and the independent variables (extracted from varstack) in the rest.
}
\description{
Prepares matrix with variables for modelling
}
\details{
Details.
}
\examples{
\dontrun{
data(biostack)
data(Oak_phylo2)
dfp <-cbind(Oak_phylo2[[1]], "pa"= rep(1,nrow(Oak_phylo2[[1]])))
dfa <-cbind(Oak_phylo2[[2]], "pa"= rep(0,nrow(Oak_phylo2[[2]])))
df3 <-rbind(dfp, dfa)
str(df3)
mat <-biomat(df3, biostack)
str(mat)
}
}
\author{
M. Iturbide \email{maibide@gmail.com}
}
\keyword{internal}

