% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdata.R
\docType{data}
\name{biostack}
\alias{biostack}
\title{Bioclim RasterStack for baseline and future data}
\description{
A RasterStack dataset containing bioclimatic variables
(bio2, bio4, bio9, bio15, bio18, bio19) for the reference period (E-OBS dataset)
and future period 1971-2100 given by (ENSEMBLES, CNRM regional climate projections).
}
\author{
M. Iturbide
}
\keyword{variables}

