% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boundingCoords.R
\name{boundingCoords}
\alias{boundingCoords}
\title{Bounding box coordinates of xy records}
\usage{
boundingCoords(xy)
}
\arguments{
\item{xy}{Data frame or list of data frames with coordinates (each row is a point)}
}
\value{
Matrix or list of matrixes with the bounding coordinates.
}
\description{
Creates the matrix of bounding coordinates around point records (xy records)
}
\examples{
\dontrun{
data(Oak_phylo2)
oak.extension <- boundingCoords(xy = Oak_phylo2)
}
}
\author{
M. Iturbide \email{maibide@gmail.com}
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutíerrez, J.M., 2015.
A framework for species distribution modelling with improved pseudo-absence generation. Ecological
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}

