% Generated by roxygen2 (4.0.2): do not edit by hand
\name{boundingCoords}
\alias{boundingCoords}
\title{Bounding box coordinates of xy records}
\usage{
boundingCoords(xy)
}
\arguments{
\item{xy}{Data frame or list of data frames with coordinates (each row is a point)}
}
\value{
Matrix or list of matrixes with the bounding coordinates.
}
\description{
Creates the matrix of bounding coordinates around point records (xy records)
}
\examples{
data(Oak_phylo2)
oak.extension <- boundingCoords(xy = Oak_phylo2)
}
\author{
M. Iturbide \email{maibide@gmail.com}
}

