% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delimit.R
\name{delimit}
\alias{delimit}
\title{Delimit study area and background}
\usage{
delimit(bounding.coords, grid, names)
}
\arguments{
\item{bounding.coords}{Object returned by function \code{\link{boundingCoords}}.
Matrix or list of matrixes of bounding coordinates indicating the maximum and minimum
values in columns and xy coordinates in rows}

\item{grid}{Projected SpatialPoints object}

\item{names}{Names or IDs to be given to each shape}
}
\value{
A list with two components:
\itemize{
 \item{bbs}{SpatialPolygons of the bounding boxes}
 \item{bbs.grid}{list(s) of matrix(ces) of the background xy coordinates in columns}.
}
}
\description{
Creation of polygon shapes from bounding coordinates and delimitation of
SpatialPoints data to the defined boundaries
}
\details{
This function is aimed at restricting the study area inside the bounding boxes
of a group of points (presences). The study area can be represented as a SpatialPoints
object of a regular point grid, this grid also represents the background (excluding
presences) from which pseudo-absences can be sampled.
Ideally, this grid is extracted from the raster objects of the variables (centroids)
to be used for modelling, see \code{\link[raster]{xyFromCell}},
\code{\link[raster]{extract}} and example below. bbs.grid is the background inside
bbs shapes.
}
\examples{
\dontrun{
data(Oak_phylo2)
data(biostack)
##creation of point grid from raster object
library(raster)
ac<-xyFromCell(biostack[[1]],  1:ncell(biostack[[1]]))
ex<-extract(biostack[[1]], ac)
##exclude sea and project
sp_grid<-SpatialPoints(ac[-which(is.na(ex)),])
projection(sp_grid)<-CRS("+proj=longlat +init=epsg:4326")
plot(sp_grid)
##delimit study area
oak.extension<-boundingCoords(Oak_phylo2)
box.grid <- delimit(bounding.coords = oak.extension, grid = sp_grid, names = names(Oak_phylo2))

## Plot presences and bounding boxes
plot(box.grid$bbs, asp = 1)
for (i in 1:length(Oak_phylo2)){
 points(Oak_phylo2[[i]], col = colors()[i*50])
 }
}
}
\author{
M. Iturbide \email{maibide@gmail.com}
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutierrez, J.M., 2015.
A framework for species distribution modelling with improved pseudo-absence generation. Ecological
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}

