% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFromPrediction.R
\name{extractFromPrediction}
\alias{extractFromPrediction}
\title{Extract values from objects or list of objects}
\usage{
extractFromPrediction(predictions, value)
}
\arguments{
\item{predictions}{listed lists of objects (e.g. as returned by \code{\link[mopa]{mopaPredict}})}

\item{value}{Character pointing to the name of component/s in the list}
}
\description{
Extrac objects from lists returned by function \code{\link[mopa]{mopaPredict}}
returned by function \code{\link[mopa]{mopaPredict}}
}
\examples{
\dontrun{
data(Oak_phylo2)
data(biostack)
projection(biostack$baseline) <- CRS("+proj=longlat +init=epsg:4326")
r <- biostack$baseline[[1]]

## Background of the whole study area
bg <- backgroundGrid(r)


## considering an unique background extent
RS_random <-pseudoAbsences(xy = Oak_phylo2, background = bg$xy, 
realizations = 10,
exclusion.buffer = 0.083*5, prevalence = -0.5, kmeans = FALSE)
fittingRS <- mopaTrain(y = RS_random, x = biostack$baseline, k = 10, 
algorithm = "glm", weighting = TRUE)

modsRS <- extractFromModel(models = fittingRS, value = "model")

#MODEL PREDICTION
prdTS.fut <- mopaPredict(models = modsTS, varstack = biostack$future)
prdTS.fut.sub <- extractFromPrediction(prdTS.fut, "realization01")
}

}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutierrez, J.M., 2015. 
A framework for species distribution modelling with improved pseudo-absence generation. Ecological 
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}
\author{
M. Iturbide
}
