% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFromPrediction.R
\name{extractFromPrediction}
\alias{extractFromPrediction}
\title{Extract values from objects or list of objects}
\usage{
extractFromPrediction(predictions, value)
}
\arguments{
\item{predictions}{Listed lists of objects (e.g. as returned by \code{\link[mopa]{mopaPredict}})}

\item{value}{Character pointing to the name of component/s in the list}
}
\description{
Extrac objects from lists returned by function \code{\link[mopa]{mopaPredict}}.
}
\examples{
\donttest{
data(Oak_phylo2)
data(biostack)
r <- biostack$baseline[[1]]
## Create background grid
bg <- backgroundGrid(r)
## Generate pseudo-absences
RS_random <-pseudoAbsences(xy = Oak_phylo2, background = bg$xy, 
                           exclusion.buffer = 0.083*5, prevalence = -0.5, kmeans = FALSE)
## Model training
fittedRS <- mopaTrain(y = RS_random, x = biostack$baseline, 
                      k = 10, algorithm = "glm", weighting = TRUE)
## Extract fitted models
mods <- extractFromModel(models = fittedRS, value = "model")
preds <- mopaPredict(models = mods, newClim = biostack$future)

## Extract predictions for the MPI regional climate model
predsMPI <- extractFromPrediction(predictions = preds, value = "MPI")
spplot(predsMPI)
}

}
\author{
M. Iturbide
}
