% Generated by roxygen2 (4.0.2): do not edit by hand
\name{indextent}
\alias{indextent}
\title{Michaelis-Mentel model fitting for extracting an index of threshold background extents}
\usage{
indextent(testmat, diagrams = FALSE)
}
\arguments{
\item{testmat}{Object returned by function loadTestValues (matrix with extents
in columns and species in rows)}

\item{diagrams}{Logical. If TRUE diagrams of the fitted Michaelis-Menten model are returned}
}
\value{
Named integer. Index of the extents at which the AUC, kappa or tss
statistics obtain the maximum score
}
\description{
Given a matrix as returned by function loadTextValues
extracts an index of the extents at which the AUC (or other test statistic) overcomes the Vm coefficient
obtained from the Michaelis-Menten model fitting.
}
\examples{
data(presaus)
data(biostack)
##modeling
modirs <-allModeling(data = presaus, varstack = biostack, k = 10, "mars")
##loading#'
auc_mars <-loadTestValues(data = presaus, "auc", "mars")

ind<-indextent(auc_mars)
}
\author{
M. Iturbide \email{maibide@gmail.com}
}

