% Generated by roxygen2 (4.0.2): do not edit by hand
\name{indextent}
\alias{indextent}
\title{Index of the extents with the maximum scores of the test statistics}
\usage{
indextent(testmat)
}
\arguments{
\item{testmat}{Object returned by function loadTestValues (matrix with extents
in columns and species in rows)}
}
\value{
Named integer. Index of the extents at which the AUC, kappa or tss
statistics obtain the maximum score
}
\description{
Given a matrix as returned by function loadTextValues
extracts an index of the extents at which the AUC, kappa or tss statistics obtain
the maximum score
}
\details{
detail.
}
\examples{
data(presaus)
data(biostack)
##modeling
modirs <-allModeling(data = presaus, varstack = biostack, k = 10, "mars")
##loading#'
auc_mars <-loadTestValues(data = presaus, "auc", "mars")

ind<-indextent(auc_mars)
}
\author{
M. Iturbide \email{maibide@gmail.com}
}

