% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maxAUCextents.R
\name{indextent}
\alias{indextent}
\title{Non-linear model fitting for extracting an index of threshold background extents}
\usage{
indextent(testmat, diagrams = TRUE)
}
\arguments{
\item{testmat}{Object returned by function loadTestValues (matrix with extents
in columns and species in rows)}

\item{diagrams}{Logical. If TRUE diagrams of the fitted models are returned}
}
\value{
Named integer. Index of the extents at which the AUC, kappa or tss
statistics obtain the maximum score. The residual sum of squares is printed in the console and
the index corresponding to the best fit is returned.
}
\description{
Given a matrix as returned by function loadTextValues
extracts an index of the extents at which the AUC (or other test statistic) overcomes the Vm coefficient
obtained from the non-linear model fitting. Non-linear asymptotic functions are tested:
Michaelis-Menten, 3-parameter asymptotic exponential (exponential 3) and 2-parameter
asymptotic exponential (exponential 2). For details about these functions, see the
"non-linear regression" section of THE R BOOK by Michael J. Crawley.
}
\examples{
\dontrun{
data(presaus)
data(biostack)
##modeling
modirs <-allModeling(data = presaus, varstack = biostack, k = 10, "mars")
##loading
auc_mars <-loadTestValues(data = presaus, "auc", "mars")
ind <- indextent(testmat = auc_mars, diagrams = TRUE)
}
}
\author{
M. Iturbide \email{maibide@gmail.com}
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutíerrez, J.M., 2015.
A framework for species distribution modelling with improved pseudo-absence generation. Ecological
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}

