% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kfold}
\alias{kfold}
\title{Stratified random partitioning into subsets}
\usage{
kfold(k = 10, df)
}
\arguments{
\item{k}{Integer. Number of subsets. Default is 10}

\item{df}{Data frame with the variable for stratification in the first column}
}
\value{
List with k data frames
}
\description{
Stratified random partitioning into subsets of presence/absence
}
\details{
Details.
}
\examples{
\dontrun{
data(biostack)
data(Oak_phylo2)
dfp <-cbind(Oak_phylo2[[1]], "pa"= rep(1,nrow(Oak_phylo2[[1]])))
dfa <-cbind(Oak_phylo2[[2]], "pa"= rep(0,nrow(Oak_phylo2[[2]])))
df3 <-rbind(dfp, dfa)
str(df3)
mat <-biomat(df3, biostack)
subsets <-kfold(10, mat)
}
}
\author{
M. Iturbide \email{maibide@gmail.com}
}
\keyword{internal}

