% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaveOneOut.R
\name{leaveOneOut}
\alias{leaveOneOut}
\title{Leave out a different subset for test each fold}
\usage{
leaveOneOut(kf.data)
}
\arguments{
\item{kf.data}{Object returned by function kfold or list of data frames}
}
\value{
List of length k. Each slot contains another list of two components:
 \item{train }{data frame of training data}
 \item{test }{data frame for test}
}
\description{
Given a data frame splitted in k subsets, this function binds 
subsets for training data and leaves one subset out for test k times, leaving 
out a different subset each fold.
}
\details{
Details.
}
\examples{
\dontrun{
data(biostack)
data(Oak_phylo2)
dfp <-cbind(Oak_phylo2[[1]], "pa"= rep(1,nrow(Oak_phylo2[[1]])))
dfa <-cbind(Oak_phylo2[[2]], "pa"= rep(0,nrow(Oak_phylo2[[2]])))
df3 <-rbind(dfp, dfa)
str(df3)
mat <- mopa:::biomat(df3, biostack$baseline)
subsets <- kfold(10, mat)
crossdata <- leaveOneOut(subsets)
}

}
\author{
M. Iturbide \email{maibide@gmail.com}
}
\keyword{internal}
