% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{loadDefinitiveModel}
\alias{loadDefinitiveModel}
\title{Load Rdata files storing definitive fitted model}
\usage{
loadDefinitiveModel(models, slot = c("allmod", "auc", "kappa", "tss", "mod",
  "p"))
}
\arguments{
\item{models}{Character of the rdata files to be loaded. Including, at least,
the attribute "species" (a character of names corresponding to the modelled species/populations).}

\item{slot}{Any character of the following: "allmod", auc", "kappa", "tss", "mod", "p"}
}
\value{
Depending on the specified slot:
 \item{allmod }{fitted model using all data for training}
 \item{auc }{AUC statistic in the cross validation}
 \item{kappa }{kappa statistic in the cross validation}
 \item{tss }{true skill statistic in the cross validation }
 \item{mod }{fitted model with partitioned data}
 \item{p }{cross model prediction}
}
\description{
Load information from Rdata generated with function allModelling based
in an index to select the definitive fitted model/s, ideally, the index returned by
function indextent should be used.
}
\details{
detail.
}
\examples{
\dontrun{
data(presaus)
data(biostackENSEMBLES)
##modeling
modirs <-allModeling(data = presaus, varstack = biostackENSEMBLES$baseline, k = 10, "mars")
##loading
auc_mars <-loadTestValues(models = modirs, test = "auc")
ind <- indextent(testmat = auc_mars, diagrams = F)
def <- loadDefinitiveModel(models = ind, slot = "allmod")
}
}
\author{
M. Iturbide
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutierrez, J.M., 2015.
A framework for species distribution modelling with improved pseudo-absence generation. Ecological
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}

