% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadDefinitiveModel.R
\name{loadDefinitiveModel}
\alias{loadDefinitiveModel}
\title{Load Rdata files storing definitive fitted model}
\usage{
loadDefinitiveModel(data, extents, slot = c("allmod", "auc", "kappa", "tss",
  "mod", "p"), algorithm = c("glm", "svm", "maxent", "mars", "randomForest",
  "cart.rpart", "cart.tree"), sourcedir = getwd())
}
\arguments{
\item{data}{Object with the same structure as the object returned by function bindPresAbs.}

\item{extents}{Named integer returned by function indextent (a named index
corresponding to the definitive extents to be considered)}

\item{slot}{Any character of the following: "allmod", auc", "kappa", "tss", "mod", "p"}

\item{algorithm}{Any character of the following: "glm", "svm", "maxent", "mars", "randomForest",
"cart.rpart" or "cart.tree"}

\item{sourcedir}{Character of the path where Rdata objects are}
}
\value{
Depending on the specified slot:
 \item{allmod }{fitted model using all data for training}
 \item{auc }{AUC statistic in the cross validation}
 \item{kappa }{kappa statistic in the cross validation}
 \item{tss }{true skill statistic in the cross validation }
 \item{mod }{fitted model with partitioned data}
 \item{p }{cross model prediction}
}
\description{
Load information from Rdata generated with function allModelling based
in an index to select the definitive fitted model/s, ideally, the index returned by
function indextent should be used.
}
\details{
detail.
}
\examples{
\dontrun{
data(presaus)
data(biostack)
##modeling
modirs <-allModeling(data = presaus, varstack = biostack, k = 10, "mars")
##loading#'
auc_mars <-loadTestValues(data = presaus, "auc", "mars")
ind <- indextent(testmat = auc_mars, diagrams = TRUE)

def <-loadDefinitiveModel(data = presaus, extents = ind, slot = "allmod", algorithm = "mars")
}
}
\author{
M. Iturbide \email{maibide@gmail.com}
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutíerrez, J.M., 2015.
A framework for species distribution modelling with improved pseudo-absence generation. Ecological
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}

