% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadTestValues.R
\name{loadTestValues}
\alias{loadTestValues}
\title{Load model validation statistics from Rdata generated with function allModelling}
\usage{
loadTestValues(data, test = c("auc", "kappa", "tss"), algorithm = c("glm",
  "svm", "maxent", "mars", "randomForest", "cart.rpart", "cart.tree"),
  sourcedir = getwd())
}
\arguments{
\item{data}{Object with the same structure as the object returned by function bindPresAbs.}

\item{test}{Any character of the following: "auc", "kappa", "tss".}

\item{algorithm}{Any character of the following: "glm", "svm", "maxent", "mars", "randomForest",
"cart.rpart" or "cart.tree"}

\item{sourcedir}{Character of the path where Rdata objects are stored}
}
\value{
Matrix with values for each species and background extents.

#
}
\description{
Load model validation statistics from Rdata generated with function allModelling
}
\details{
detail.
}
\examples{
\dontrun{
data(presaus)
data(biostack)
##modeling
modirs <-allModeling(data = presaus, varstack = biostack, k = 10, "mars")
##loading
auc_mars <-loadTestValues(data = presaus, test = "auc", algorithm = "mars")

library(lattice)
levelplot(auc_mars, aspect = 5,
          scales = list(y = list(cex = 0.8,
                                 at = c(1, 49, ncol(auc_mars)),
                                 labels = c(colnames(auc_mars)[1],
                                            colnames(auc_mars)[49],
                                            colnames(auc_mars)[ncol(auc_mars)]))),
          at = seq(0.6, 1, 0.01), col.regions = bpy.colors,
          xlab = "Haplogroups", ylab = "Background extent (km)", main = "AUC")
}
}
\author{
M. Iturbide \email{maibide@gmail.com}
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutíerrez, J.M., 2015.
A framework for species distribution modelling with improved pseudo-absence generation. Ecological
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}

