% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadTestValues.R
\name{loadTestValues}
\alias{loadTestValues}
\title{Load model validation statistics from Rdata generated with function allModelling}
\usage{
loadTestValues(models, sourcedir = getwd(), test = c("auc", "kappa", "tss"))
}
\arguments{
\item{models}{Object with the same structure as the object returned by function allModelling.}

\item{sourcedir}{Directory of the objects generated and saved by allModelling}

\item{test}{Any character of the following: "auc", "kappa", "tss".
"cart.rpart" or "cart.tree"}
}
\value{
Matrix with values for each species and background extents.

#
}
\description{
Load model validation statistics from Rdata generated with function allModelling
}
\details{
detail.
}
\examples{
\dontrun{
data(presaus)
data(biostackENSEMBLES)
##modeling
modirs <-allModeling(data = presaus, varstack = biostackENSEMBLES$baseline, k = 10, "mars")
##loading
auc_mars <-loadTestValues(models = modirs, test = "auc")

library(lattice)
levelplot(auc_mars, aspect = 5,
          scales = list(y = list(cex = 0.8,
                                 at = c(1, 49, ncol(auc_mars)),
                                 labels = c(colnames(auc_mars)[1],
                                            colnames(auc_mars)[49],
                                            colnames(auc_mars)[ncol(auc_mars)]))),
          at = seq(0.6, 1, 0.01), col.regions = bpy.colors,
          xlab = "Haplogroups", ylab = "Background extent (km)", main = "AUC")
}
}
\author{
M. Iturbide
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutierrez, J.M., 2015.
A framework for species distribution modelling with improved pseudo-absence generation. Ecological
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}

