% Generated by roxygen2 (4.0.2): do not edit by hand
\name{modelo}
\alias{modelo}
\title{Species distribution modelling and cross validation}
\usage{
modelo(kdata, data, algorithm = c("glm", "svm", "maxent", "mars",
  "randomForest", "cart.rpart", "cart.tree"))
}
\arguments{
\item{kdata}{Object returned by function leaveOneOut}

\item{data}{Object returned by function biomat. 2D matrix with the dependent variable
(presence/absence) in the first column and the independent variables in the rest
(extracted from varstack)}

\item{algorithm}{Any character of the following: "glm", "svm", "maxent", "mars", "randomForest", "cart.rpart"
or "cart.tree"}
}
\value{
A list with six components:
 \item{allmod }{fitted model using all data for training}
 \item{auc }{AUC statistic in the cross validation}
 \item{kappa }{kappa statistic in the cross validation}
 \item{tss }{true skill statistic in the cross validation }
 \item{mod }{fitted model with partitioned data}
 \item{p }{cross model prediction}
}
\description{
Species distribution modelling with k-fold cross validation.
Algorithms supported are "glm", "svm", "maxent", "mars", "randomForest", "cart.rpart"
and "cart.tree"
}
\details{
This function calculates the AUC with the function “auc” from package
“PresenceAbsence”. Package SDMTools must be detached.
}
\examples{
##SpatialPoints grid
data(sp_grid)
class(sp_grid)
}
\author{
M. Iturbide \email{maibide@gmail.com}
}
\keyword{internal}

