% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mopaPredict.R
\name{mopaPredict}
\alias{mopaPredict}
\title{Model prediction}
\usage{
mopaPredict(models, newClim)
}
\arguments{
\item{models}{model class object (e.g. "glm") or list of model class objects, e.g. as returned by function \code{\link[mopa]{extractFromModel}}.}

\item{newClim}{RasterStack or list of RasterStack objects with variables for projecting}
}
\value{
RasterStack of the projected probabilities
}
\description{
Model projection into a RasterStack
}
\examples{
data(Oak_phylo2)
data(biostack)
projection(biostack$baseline) <- CRS("+proj=longlat +init=epsg:4326")
r <- biostack$baseline[[1]]

## Background of the whole study area
bg <- backgroundGrid(r)

## inside different background extents
bg.extents <- backgroundRadius(xy = Oak_phylo2, background = bg$xy, 
start = 0.166, by = 0.083*10, unit = "decimal degrees")
TS_random <-pseudoAbsences(xy = Oak_phylo2, background = bg.extents, 
exclusion.buffer = 0.083*5, prevalence = -0.5, kmeans = FALSE)
fittingTS <- mopaTrain(y = TS_random, x = biostack$baseline, k = 10, 
algorithm = "glm", weighting = TRUE, diagrams = TRUE)

## considering an unique background extent
RS_random <-pseudoAbsences(xy = Oak_phylo2, background = bg$xy, 
exclusion.buffer = 0.083*5, prevalence = -0.5, kmeans = FALSE)
fittingRS <- mopaTrain(y = RS_random, x = biostack$baseline, k = 10, 
algorithm = "glm", weighting = TRUE)

modsTS <- extractFromModel(models = fittingTS, value = "model")

#MODEL PREDICTION
prdTS <- mopaPredict(models = modsTS, newClim = biostack$baseline)
spplot(stack(prdTS$H11$PA01$glm$baseClim01))
prdTS.fut <- mopaPredict(models = modsTS, newClim = biostack$future)
spplot(stack(prdTS.fut$H11$PA01$glm$baseClim01$MPI))

}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutierrez, J.M., 2015. 
A framework for species distribution modelling with improved pseudo-absence generation. Ecological 
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}
\seealso{
\code{\link[mopa]{mopaTrain}}
}
\author{
M. Iturbide
}
