% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mopaPredict.R
\name{mopaPredict0}
\alias{mopaPredict0}
\title{Model prediction}
\usage{
mopaPredict0(models, varstack)
}
\arguments{
\item{models}{model class object (e.g. "glm") or list of model class objects, e.g. as returned by function \code{\link[mopa]{extractFromModel}}.}

\item{varstack}{RasterStack or list of RasterStacks of variables for projecting. If list, named lists are
recommended}
}
\value{
RasterStack of the projected probabilities
}
\description{
Model projection into a RasterStack
}
\author{
M. Iturbide
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutierrez, J.M., 2015.
A framework for species distribution modelling with improved pseudo-absence generation. Ecological
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}
\seealso{
\code{\link[mopa]{mopaTrain}}
}

