% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdata.R
\docType{data}
\name{Oak_phylo2}
\alias{Oak_phylo2}
\title{Oak distribution}
\source{
Oak_phylo2 is a modified subset of the \strong{Quercus sp Europe Petit 2002}
database (Petit et al., 2002b), which is available in the \strong{Georeferenced Database
of Genetic Diversity} or \strong{(GD)^2}.
}
\description{
A dataset consisting of a list with two
data frames with xy coordinates corresponding to the distribution of two
phylogenetic groups of oaks (H7 and H10)
}
\references{
Petit, R. J. \emph{et al} 2002. Chloroplast DNA variation in european
white oaks: Phylogeography and patterns of diversity based on data from over 2600 populations.
Forest Ecology and Management 156 (1-3), 5-26.
}
\keyword{records}
\keyword{xy}

