% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdata.R
\docType{data}
\name{presaus}
\alias{presaus}
\title{Presence/Pseudo-absence data for different background extents}
\description{
A dataset with listed data frames for two different Oak groups (H11 and H5).
Each data frame is the distribution of presences and pseudo-absences inside
a certain background extent
}
\author{
M. Iturbide
}
\keyword{background}

