% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseudoAbsences.R
\name{pseudoAbsences}
\alias{pseudoAbsences}
\title{Pseudo-absences}
\usage{
pseudoAbsences(xy, background, realizations = 1, exclusion.buffer = 0.0166,
  prevalence = 0.5, kmeans = FALSE, varstack = NULL)
}
\arguments{
\item{xy}{Data frame or list of data frames with coordinates (each row is a point)}

\item{background}{Matrix or list/s of matrixes with background coordinates in columns.
Object derived from function delimit, OCSVMprofiling or bgRadio.}

\item{realizations}{Integer. Number of realizations (default = 1).}

\item{exclusion.buffer}{value of the minimum distance to be kept between presence data and
pseudo-absence data. Default is 0.0166}

\item{prevalence}{Proportion of presences against absences. Default is 0.5 (equal number of
presences and absences)}

\item{kmeans}{Logical. If FALSE (default) pseudo-absences are generated at random. If TRUE
k-means clustering of the background is done and centroids are extracted as pseudo-absences.}

\item{varstack}{RasterStack of variables for modelling}
}
\value{
List/s of data frames
}
\description{
Pseudo-absence data generation at random or by k-means clustering inside backgrounds of
different extent
}
\details{
Details.The application of this function is the third step in a three-step proccess
to generate pseudo-absences.
}
\examples{
\dontrun{
data(Oak_phylo2)
data(biostack)
projection(biostack$baseline) <- CRS("+proj=longlat +init=epsg:4326")
r <- biostack$baseline[[1]]
## Background of the whole study area
bg <- backgroundGrid(r)

## environmental profiling
bg.profiled <- OCSVMprofiling(xy = Oak_phylo2, varstack = biostack$baseline,
background = bg$xy)

## different background extents
bg.extents <- backgroundRadios(xy = Oak_phylo2, background = bg$xy,
start = 0.166, by = 0.083*10, unit = "decimal degrees")
bg.extents2 <- backgroundRadios(xy = Oak_phylo2, background = bg.profiled$absence,
start = 0.166, by = 0.083*10, unit = "decimal degrees")

## inside different background extents
TS_random <-pseudoAbsences(xy = Oak_phylo2, background = bg.extents2,
exclusion.buffer = 0.083*5, prevalence = -0.5, kmeans = FALSE)

## considering an unique background extent
RS_random <-pseudoAbsences(xy = Oak_phylo2, background = bg$xy,
exclusion.buffer = 0.083*5, prevalence = -0.5, kmeans = FALSE)
RSEP_random <-pseudoAbsences(xy = Oak_phylo2, background = bg.profiled$absence,
exclusion.buffer = 0.083*5, prevalence = -0.5, kmeans = FALSE)

## with k-means clustering
TS_kmeans <-pseudoAbsences(xy = Oak_phylo2, background = bg.extents2,
exclusion.buffer = 0.083*5, prevalence = -0.5, kmeans = TRUE, varstack = biostack$baseline)
}
}
\author{
M. Iturbide
}
\references{
Iturbide, M., Bedia, J., Herrera, S., del Hierro, O., Pinto, M., Gutierrez, J.M., 2015.
A framework for species distribution modelling with improved pseudo-absence generation. Ecological
Modelling. DOI:10.1016/j.ecolmodel.2015.05.018.
}
\seealso{
\code{\link[mopa]{mopaTrain}}
}

